SortMyFolder/requirements.txt
openai
python-dotenv
Pillow
PyPDF2
